requirements.txt
# Real Madrid 2023-2024 Match Analysis System Dependencies

# Database connectivity
psycopg2-binary==2.9.7

# Configuration and environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Data analysis and processing
pandas==2.1.1
numpy==1.24.3

# API requests and web scraping
requests==2.31.0
urllib3==2.0.4

# Web framework for interface
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
jinja2==3.1.2
werkzeug==2.3.7

# Development and analysis tools
jupyter==1.0.0
ipython==8.15.0
matplotlib==3.7.2
seaborn==0.12.2

# Optional: Advanced analytics
scikit-learn==1.3.0
plotly==5.15.0

# Caching (optional)
redis==4.6.0

# Logging and utilities
colorama==0.4.6
